Adds supports for kinematicBodies and free form movement. Allows for more advanced physics

+++ components/PositionComponent.js
@@ -2,8 +2,8 @@ import {defineComponent, Types} from "../bitecs.mjs";
 
 
 export const PositionComponent = defineComponent({
-        x: Types.i32,
-        y: Types.i32
+        x: Types.f32,
+        y: Types.f32
 });
 
 export default PositionComponent;
